[N64] PI DMA fix and temporary workaround
Temporary workaround to make nm64 work. If cart and rdram are both in LE, making a simple dma to rdram does not work if either addr is not work-aligned. E.g.: given cart addr 0x0032e6a6 and aligned rdram:
- cart BE, rdram BE
  cart 0x0032e6a4 : ffff0000 00120000 00100028
  rdram                 0000 00120000 00100028
- cart LE, rdram LE
  cart 0x0032e6a4 : 0000ffff 00001200 28001000
  rdram           :     ffff 00001200 28001000
The two first two bytes copied are incorrect. Temporarily, make cart BE and use the below..